Added barrelling for repeat button
